requirements.txt
jupyter==1.0.0
numpy==1.19.4
pandas==1.2.0
pyro-ppl==1.5.1
pytorch-lightning==1.1.2
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
tensorboard==2.4.0
torch==1.7.1
torchvision==0.8.2